requirements.txt
streamlit
streamlit-chat
openai
python-dotenv
tiktoken
langchain